Create Data Management App
This commit helps condencse the octoprint stuff and put it all in the same place
