+++ forge/admin.py
@@ -2,6 +2,7 @@ from django.contrib import admin
 
 from machine_management.models import *
 from business.models import *
+from data_management.models import *
 from user_management.models import *
 from apis.models import *
