+++ forge/apps.py
@@ -32,4 +32,10 @@ class APIConfig(AppConfig):
     name = 'apis'
     
 class BusinessConfig(AppConfig):
+    # ! function: App that handles all buisness operations with  users
     name = 'business'
+
+
+class DataManagementConfig(AppConfig):
+    # ! function: App that handles all interactions with octoprint information
+    name = 'data_management'
